Use 'nginx -T' to dump active config lines
It could be that using "find -name *.conf" could find old non-active
files, which are not relevant to the live config. For example, if
you took an old copy of files in "backup_dir/*.conf".

Using nginx -T is similar in spirit to how the apache-based certwatch
uses "httpd -t -DDUMP_CERTS", but you still have to grep out the
lines you want.